--- src/Shopway.Application/Features/Proxy/GenericPageQuery/GenericPageQueryDiscriminator.cs
@@ -1,5 +1,0 @@
-﻿using Shopway.Domain.Common.Discriminators;
-
-namespace Shopway.Application.Features.Proxy;
-
-public sealed record class GenericPageQueryDiscriminator(string Entity, Type PageType) : Discriminator;
