--- src/Shopway.Application/Features/Proxy/GenericQuery/QueryById/GenericByIdQueryDiscriminator.cs
@@ -1,5 +1,0 @@
-﻿using Shopway.Domain.Common.Discriminators;
-
-namespace Shopway.Application.Features.Proxy.GenericQuery.QueryById;
-
-public sealed record class GenericByIdQueryDiscriminator(string Entity) : Discriminator;
