--- src/Shopway.Application/Features/Proxy/GenericQuery/QueryByKey/GenericByKeyQueryDiscriminator.cs
@@ -1,5 +1,0 @@
-﻿using Shopway.Domain.Common.Discriminators;
-
-namespace Shopway.Application.Features.Proxy.GenericQuery.QueryByKey;
-
-public sealed record class GenericByKeyQueryDiscriminator(string Entity) : Discriminator;
