+++ src/Shopway.Presentation/Controllers/ProxyController.cs
@@ -6,7 +6,7 @@ using Microsoft.AspNetCore.Mvc;
 using Shopway.Application.Abstractions.CQRS;
 using Shopway.Application.Features;
 using Shopway.Application.Features.Proxy;
-using Shopway.Application.Features.Proxy.GenericQuery.QueryByKey;
+using Shopway.Application.Features.Proxy.GenericQuery;
 using Shopway.Application.Features.Proxy.PageQuery;
 using Shopway.Application.Features.Proxy.Query;
 using Shopway.Presentation.Abstractions;
